pyproject.toml
[tool.poetry]
name = "niyo-task-mgmt-system"
version = "1.0.0"
description = "Niyo Task Management System"
authors = ["Osayi Akoko <osayistreams@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
django-split-settings = "^1.3.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
user-agents = "^2.2.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
pytest-asyncio = "^0.23.7"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest-django = "^4.8.0"
pytest-xdist = "^3.6.1"
model-bakery = "^1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
django_find_project = false
python_files = ["**/test*.py", ]
pythonpath = "./src"